fix ordering of standard out/error messages
Fixes a race condition bug where events can get written to the terminal out of order if standard out & standard error are used around the same time.

I also removed the level filtering on event handlers. The only use case was the multiple handlers for stdout/stderr, which is no longer a use case.
If level filtering is to be done, it should be done within the handler.

closes #2