add better support for error interface
If a struct was passed, and the struct satisfied the `error` interface, but didn't have a `String()` method, we'd treat it as an emptry struct.
This change makes it do what is expected. Basically we get the error value instead. However, if the struct has exported attributes and satisfies the `error` interface, we add a `Error` field instead.