remove any possibility of a race from DefaultLogger()
previously there was a potential race condition that might occur during the very first call to DefaultLogger(). The race would result in 2 *Logger objects being created, and one of them ending up getting garbage collected.
This is a minor issue, that would have been extremely hard to trigger, but it was solved to make the golang race detector shut up.